+++ svm-validation-interval/src/main/java/ru/spbftu/igorbotian/phdapp/svm/validation/PreciseIntervalJudgementsRatioAnalyzer.java
@@ -52,7 +52,8 @@ class PreciseIntervalJudgementsRatioAnalyzer
                     specificClassifierParams,
                     override(specificValidatorParams, Collections.singleton(ratioParam))
             ));
-            fireCrossValidationContinued(100 * ((i - lowerBound) / (upperBound - lowerBound)));
+            fireCrossValidationContinued((int) (100 * (((float) i - (float) lowerBound)
+                    / ((float) upperBound - (float) lowerBound))));
 
             if (processInterrupted()) {
                 fireCrossValidationInterrupted();
